requirements.txt
Django>=4.2.13,<5
pytz==2024.1
httplib2==0.22.0
feedparser==6.0.11
Markdown==3.6
psycopg==3.1.19
versiontools==1.9.1
statsd==4.0.1
pep8==1.7.1
pyflakes==3.2.0
mccabe==0.7.0
configparser==7.0.0
pycodestyle==2.11.1
importlib-metadata<7.2 # for flake8
zipp==3.18.2  # for flake8
flake8==7.0.0
tornado==6.4
beautifulsoup4==4.12.3
cssselect==1.2.0
lxml==5.2.2
fuzzywuzzy==0.18.0
sure==2.0.1
coverage==7.5.1
pyasn1==0.6.0
cffi==1.16.0  # pyOpenSSL
cryptography==42.0.7  # pyOpenSSL
pyOpenSSL==24.1.0
ndg-httpsclient==0.5.1
urllib3==2.2.1  # requests, sentry-sdk
chardet==5.2.0  # requests
idna==3.7  # requests
certifi==2024.2.2  # requests, sentry-sdk
requests==2.32.2
logilab-common==2.0.0
astroid==3.2.2
pylint==3.2.2
six==1.16.0
ipaddress==1.0.23  # faker
faker==25.2.0  # factory_boy
factory_boy==3.3.0

ldap3==2.9.1
sqlparse==0.5.0

s3transfer==0.10.1
jmespath==1.0.1

boto3>=1.12.0,<1.35.0
botocore>=1.15.0,<1.35.0

python-dateutil==2.9.0.post0

django-indexer==0.3.0
django-templatetag-sugar==1.0
django-paging==0.2.5
django-annoying==0.10.6
django-appconf==1.0.6
django-statsd-mozilla==0.4.0
raven==6.10.0
sentry-sdk==2.2.0
django-bootstrap4==24.3
django-debug-toolbar==4.3.0

django-smoketest==1.2.1
# For django-extensions
django-extensions==3.2.3
django-stagingcontext==0.1.0
django-ga-context==0.1.0
django-impersonate==1.9.2
gunicorn==22.0.0
django-flatblocks==1.0.0
django-storages==1.14.3

djangorestframework==3.15.1

ctlsettings==0.3.2

text_unidecode==1.3

django-taggit==5.0.1

pbr==6.0.0 # bandit
pyYAML==6.0.1 # bandit
stevedore==5.2.0  # bandit
rich==13.7.1 # bandit
bandit==1.7.8

django-smtp-ssl==1.0
django-contact-us==1.1.0

entrypoints==0.4
distro==1.9.0
typing_extensions==4.11.0

python-cas==1.6.0
django-cas-ng==5.0.1
asgiref==3.8.1

backports.zoneinfo;python_version<"3.9"
